// src/api/models/CategoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CategoryDto = {
    id: string;
    key: string;
    label: string;
    description?: Record<string, any> | null;
    enabled: boolean;
    isDefault?: boolean;
    sort: number;
    createdAt: string;
    updatedAt: string;
};